+++ client/src/screens/hub/HubMainScreen/components/ControlsSidebar/components/ControlsSidebarXAlignment/ControlsSidebarXAlignment.js
@@ -39,6 +39,8 @@ function ControlsSidebarXAlignment(props) {
   useEffect(() => {
     if (opened && popupRef.current) {
       popupRef.current.focus();
+      const { top } = popupRef.current.getBoundingClientRect();
+      popupRef.current.style.maxHeight = `${window.innerHeight - top - 10}px`;
     }
   }, [opened]);