+++ client/src/screens/hub/HubMainScreen/components/ControlsSidebar/components/GroupByStyle/GroupByStyle.js
@@ -29,7 +29,11 @@ function GroupByStyle(props) {
 
   useEffect(() => {
     if (opened) {
-      popupRef.current?.focus();
+      if (popupRef.current) {
+        popupRef.current.focus();
+        const { top } = popupRef.current.getBoundingClientRect();
+        popupRef.current.style.maxHeight = `${window.innerHeight - top - 10}px`;
+      }
       dropdownRef.current?.selectRef?.current?.focus();
     }
   }, [opened]);